backend/main.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
import socketio
import uvicorn
import asyncio
import os
import random
import math
import urllib.request
import json
from dotenv import load_dotenv
from database import init_db

load_dotenv()

app = FastAPI(title="AutoRoute Intelligence Agent Core")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

sio = socketio.AsyncServer(async_mode='asgi', cors_allowed_origins='*')
socket_app = socketio.ASGIApp(sio, other_asgi_app=app)

simulation_task = None
gps_jammer_active = False

driver_names = [
    "Muthusamy K.", "Karthik R.", "Saravanan T.", "Vignesh P.", "Senthil Kumar",
    "Rajesh K.", "Prabhu M.", "Prakash S.", "Manikandan V.", "Gopinath L.",
    "Tamilselvan R.", "Balaji S.", "Arunachalam M.", "Dinesh Kumar", "Kannan P.",
    "Velmurugan T.", "Sivakumar K.", "Ganesan N.", "Mahesh R.", "Subramanian V."
]

conductor_names = [
    "Ramesh P.", "Suresh V.", "Ajith Kumar", "Selvam N.", "Murugan K.",
    "Pandian M.", "Kumaravel R.", "Jayakumar S.", "Palanisamy T.", "Arumugam V.",
    "Iyappan S.", "Krishnan L.", "Balamurugan M.", "Ramakrishnan K.", "Nagarajan P.",
    "Srinivasan V.", "Anbuselvan R.", "Chandran M.", "Mohan Raj", "Sathish Kumar"
]

# True coordinates matched precisely with JavaScript Frontend logic
LANDMARKS = {
    "Central Bus Stand": (10.9600, 78.0750),
    "Railway Station": (10.9650, 78.0800),
    "Pasupathipalayam": (10.9500, 78.0700),
    "Gandhigramam": (10.9550, 78.0850),
    "Vengamedu": (10.9700, 78.0900),
    "Thanthonimalai": (10.9400, 78.0650),
    "Vangal Road": (10.9800, 78.0600),
}

routes_def = {
    "R1 (Thanthonimalai-South)": ["Thanthonimalai", "Pasupathipalayam", "Thanthonimalai"],
    "R2 (East-Gandhigramam)": ["Gandhigramam", "Railway Station", "Gandhigramam"],
    "R3 (North-Vangal)": ["Vangal Road", "Vengamedu", "Vangal Road"],
    "R4 (Central-Core)": ["Central Bus Stand", "Railway Station", "Pasupathipalayam", "Central Bus Stand"],
    "R5 (Outer Loop)": ["Thanthonimalai", "Gandhigramam", "Vengamedu", "Thanthonimalai"]
}

def create_route_waypoints(stops):
    import urllib.request, urllib.error, json
    full_path = []
    
    # Attempt to strictly trace exact road geometries using OSRM
    coords_str = ";".join([f"{s[1]},{s[0]}" for s in stops])
    url = f"http://router.project-osrm.org/route/v1/driving/{coords_str}?geometries=geojson&overview=full"
    
    try:
        req = urllib.request.Request(url, headers={'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)'})
        with urllib.request.urlopen(req, timeout=10) as response:
            data = json.loads(response.read().decode())
            if data.get("code") == "Ok":
                coords = data['routes'][0]['geometry']['coordinates']
                for c in coords:
                    full_path.append([c[1], c[0]])
                return full_path
    except Exception as e:
        print(f"OSRM Physical Snapping Blocked: {e}")

    print("Falling back to absolute geometric interpolation.")
    for i in range(len(stops)-1):
        s1 = stops[i]
        s2 = stops[i+1]
        dist = math.sqrt((s2[0]-s1[0])**2 + (s2[1]-s1[1])**2)
        steps = max(int(dist / 0.0003), 40) 
        for step in range(steps):
            lat = s1[0] + (s2[0]-s1[0])*(step/steps)
            lng = s1[1] + (s2[1]-s1[1])*(step/steps)
            full_path.append([lat, lng])
    full_path.append([stops[-1][0], stops[-1][1]])
    return full_path

route_geometries = {}
for r_name, j_seq in routes_def.items():
    stops = [LANDMARKS[j] for j in j_seq]
    route_geometries[r_name] = create_route_waypoints(stops)

BUSES_DATA = {}
for i in range(20):
    bus_id = f"BUS_{i+1:02d}"
    
    # Mathematically distribute evenly so buses DO NOT bundle up at center
    rt_keys = list(routes_def.keys())
    rt_name = rt_keys[i % len(rt_keys)]
    geom = route_geometries[rt_name]
    
    # Calculate staggered starting index
    segment_spacing = len(geom) // 4
    start_idx = (segment_spacing * (i // len(rt_keys))) % len(geom)
    start_pt = geom[start_idx]
    
    BUSES_DATA[bus_id] = {
        "id": bus_id,
        "driver": driver_names[i],
        "conductor": conductor_names[i],
        "route": rt_name,
        "route_points": geom,
        "path_index": start_idx,
        "status": "active",
        "lat": start_pt[0],
        "lng": start_pt[1],
        "speed": random.randint(15, 30), # Reduced physical speed for smooth stable tracing
        "heading": 0,
        "fuel": random.randint(30, 95),
        "tickets": random.randint(10, 200),
        "traffic_delay": 0
    }

@app.post("/trigger_fault")
async def trigger_fault_organic():
    active = [b['id'] for b in BUSES_DATA.values() if b['status'] == "active"]
    if len(active) < 2: return {"error": "not enough active"}
    bus_id = random.choice(active)
    BUSES_DATA[bus_id]["status"] = "offline"
    available_buses = [b for b in BUSES_DATA.values() if b["status"] == "active" and b["id"] != bus_id]
    if not available_buses: return {"error": "no backups"}
    backup_bus = min(available_buses, key=lambda k: k['tickets'])
    decision = {
        "event_type": "FAULT_DETECTED",
        "title": f"XAI REASONING: FAULT ON {bus_id}",
        "reasoning": f"[EXPLAINABLE AI]\n> Primary Alert: Hardware fault detected on {bus_id}.\n> Constraint Resolution: Locating lowest yield asset to minimize revenue loss.\n> AI Confidence: 98.4%\n> Directed Action: Rerouting {backup_bus['id']} (Current Yield: {backup_bus['tickets']}) to assist {bus_id}.",
        "action": f"REROUTE_{backup_bus['id']}_TO_ASSIST",
        "affected_buses": [bus_id, backup_bus["id"]]
    }
    await sio.emit("agent_alert", decision)
    return decision

@app.post("/trigger_event")
async def trigger_surge_organic():
    underperf = sorted([b for b in BUSES_DATA.values() if b["status"] == "active"], key=lambda x: x["tickets"])[:2]
    bus_ids = [b["id"] for b in underperf]
    if len(bus_ids) < 2: return {"error": "not enough buses"}
    decision = {
        "event_type": "CROWD_SURGE",
        "title": "XAI REASONING: CROWD SURGE ALARM",
        "reasoning": f"[EXPLAINABLE AI]\n> Environment Trigger: Extreme optical density tracking in Sector A.\n> Operational Logic: Identified lowest-performing assets ({', '.join(bus_ids)}).\n> AI Confidence: 93.1%\n> Directed Action: Redeploying vectors to bottleneck zone for rapid surge clearance.",
        "action": "DEPLOY_INTERVENTION_SQUAD",
        "affected_buses": bus_ids
    }
    await sio.emit("agent_alert", decision)
    return decision

@app.post("/trigger_rebalance")
async def trigger_rebalance_organic():
    active = [b['id'] for b in BUSES_DATA.values() if b["status"] == "active"]
    if len(active) < 2: return {"error": "not enough buses"}
    targets = random.sample(active, 2)
    decision = {
        "event_type": "EFFICIENCY_OPTIMIZATION",
        "title": "XAI REASONING: DYNAMIC REBALANCE",
        "reasoning": f"[EXPLAINABLE AI]\n> Analytics Engine: Corridor ticket velocities highly skewed.\n> Resource Analysis: Extracting {targets[0]} from saturated network, shifting to underserved routes.\n> Predictive Lift: Expected +15% fleet revenue efficiency.\n> AI Confidence: 91.0%",
        "action": "DYNAMIC_CORRIDOR_SHIFT",
        "affected_buses": targets
    }
    await sio.emit("agent_alert", decision)
    return decision

@app.post("/trigger_predictive")
async def trigger_predictive():
    decision = {
        "event_type": "PREDICTIVE_INTELLIGENCE",
        "title": "XAI REASONING: TEMPORAL DEMAND",
        "reasoning": "[EXPLAINABLE AI]\n> Deep Learning Forecast: Event history indicates Friday 17:00 Railway demand surge (Probability: 88%).\n> Proactive Logic: Pre-positioning free assets immediately to minimize transit latency.\n> Directed Action: Shifting BUS_05 to holding pattern near terminal.",
        "action": "TACTICAL_PRE_POSITION",
        "affected_buses": ["BUS_05"]
    }
    await sio.emit("agent_alert", decision)
    return decision

@app.post("/toggle_gps_jammer")
async def toggle_gps_jammer():
    global gps_jammer_active
    gps_jammer_active = not gps_jammer_active
    return {"status": "GPS Jammer Active" if gps_jammer_active else "GPS Jammer Disabled"}

@app.on_event("startup")
async def startup_event():
    init_db()
    global simulation_task
    simulation_task = asyncio.create_task(simulation_loop())
    print("Production Agent Engine Started (Automated Organic XAI Events Enabled).")

@sio.event
async def connect(sid, environ):
    await sio.emit("bus_positions", {"buses": list(BUSES_DATA.values())}, to=sid)

async def simulation_loop():
    while True:
        await asyncio.sleep(1.0)
        
        # Organic automated events to completely remove "Simulator" dependency
        if random.random() > 0.990:
            event_type = random.choice(["fault", "surge", "rebalance"])
            if event_type == "fault": await trigger_fault_organic()
            elif event_type == "surge": await trigger_surge_organic()
            elif event_type == "rebalance": await trigger_rebalance_organic()
                
        for bus_id, bus in BUSES_DATA.items():
            if bus["status"] in ["active", "assisting"]:
                if bus["traffic_delay"] > 0:
                    bus['speed'] = random.uniform(0, 5)
                    bus['traffic_delay'] -= 1
                    continue
                    
                pts = bus['route_points']
                idx = bus['path_index']
                
                if idx >= len(pts) - 1:
                    bus['path_index'] = 0
                    idx = 0
                    
                target = pts[idx + 1]
                current = [bus['lat'], bus['lng']]
                
                dist = math.sqrt((target[0]-current[0])**2 + (target[1]-current[1])**2)
                step_size = (bus['speed'] / 3.6) * 1.0 / 111320.0
                
                if dist <= step_size:
                    bus['path_index'] = idx + 1
                    bus['lat'], bus['lng'] = target[0], target[1]
                    
                    if random.random() > 0.96:
                        bus['traffic_delay'] = random.randint(8, 15) 
                else:
                    ratio = step_size / dist if dist != 0 else 0
                    bus['lat'] += (target[0]-current[0]) * ratio
                    bus['lng'] += (target[1]-current[1]) * ratio
                    bus['heading'] = math.degrees(math.atan2(target[1]-current[1], target[0]-current[0]))
                
                target_spd = random.randint(15, 30)
                bus['speed'] += (target_spd - bus['speed']) * 0.15
                
                if random.random() > 0.8: bus['tickets'] += random.randint(1, 4)
                
                # Slower fuel drain AND massive automatic refuel guard logic
                if random.random() > 0.98: 
                    bus['fuel'] -= 1
                    if bus['fuel'] <= 0:
                        bus['fuel'] = 100
                        
        clean_data = [ {k: v for k, v in b.items() if k not in ["route_points", "path_index"]} for b in BUSES_DATA.values() ]
        await sio.emit("bus_positions", {"buses": clean_data})
