backend/requirements.txt
fastapi
uvicorn
python-socketio
groq
mysql-connector-python
python-dotenv
